+++ actionpack/lib/action_dispatch/journey/route.rb
@@ -1,7 +1,7 @@
 module ActionDispatch
   module Journey # :nodoc:
     class Route # :nodoc:
-      attr_reader :app, :path, :verb, :defaults, :ip, :name
+      attr_reader :app, :path, :defaults, :name
 
       attr_reader :constraints
       alias :conditions :constraints
@@ -12,15 +12,11 @@ module ActionDispatch
       # +path+ is a path constraint.
       # +constraints+ is a hash of constraints to be applied to this route.
       def initialize(name, app, path, constraints, defaults = {})
-        constraints  = constraints.dup
         @name        = name
         @app         = app
         @path        = path
-        @verb        = constraints[:request_method] || //
-        @ip          = constraints.delete(:ip) || //
 
         @constraints = constraints
-        @constraints.keep_if { |_,v| Regexp === v || String === v }
         @defaults    = defaults
         @required_defaults = nil
         @required_parts    = nil
@@ -88,6 +84,29 @@ module ActionDispatch
           matches = parts
           @defaults.dup.delete_if { |k,_| matches.include?(k) }
         end
+      end
+
+      def matches?(request)
+        constraints.all? do |method, value|
+          next true unless request.respond_to?(method)
+
+          case value
+          when Regexp, String
+            value === request.send(method).to_s
+          when Array
+            value.include?(request.send(method))
+          else
+            value === request.send(method)
+          end
+        end
+      end
+
+      def ip
+        constraints[:ip] || //
+      end
+
+      def verb
+        constraints[:request_method] || //
       end
     end
   end